pyproject.toml
[project]
name = "bonsai-prp"
dynamic = ["version"]
authors = [
  {name = "Markus Johansson", email = "markus.h.johansson@skane.se"},
  {name = "Ryan Kennedy", email = "Ryan.Kennedy@skane.se"},
]
description = "Pipline result processing program for the JASEN pipeline and Bonsai tool."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.4,<3",
    "pyyaml>=6,<7",
    "packaging>=25",
]

[project.urls]
Repository = "https://github.com/Clinical-Genomics-Lund/"
Issues = "https://github.com/Clinical-Genomics-Lund/bonsai-prp/issues"
Changelog = "https://github.com/Clinical-Genomics-Lund/CHANGELOG.md"

[project.scripts]
prp = "prp.cli:cli"

[project.optional-dependencies]


# upload = cli + requests
upload = [
    "bonsai-libs @ git+https://github.com/mhkc/bonsai-libs.git@v0.1.0",
    "click>=8.1",
    "requests>=2.31",
]

# analysis = cli + all heavy tools
analysis = [
    "click>=8.1",
    "numpy>=1.26,<2.0",
    "pandas>=2.1,<2.3",
    "biopython>=1.83,<1.86",
    "cyvcf2>=0.31,<0.32",
    "pysam>=0.22,<0.23",
]

all = [
    "bonsai-libs @ git+https://github.com/mhkc/bonsai-libs.git@v0.1.0",
    "click>=8.1",
    "requests>=2.31,<3",
    "numpy>=1.26,<2.0",
    "pandas>=2.1,<2.3",
    "biopython>=1.83,<1.86",
    "cyvcf2>=0.31,<0.32",
    "pysam>=0.22,<0.23",
]

# dev-only tools (not installed with 'all')
dev = [
    "pylint>=3.0",
    "black>=23.11",
    "isort>=5.12",
]

# test-only tools (not installed with 'all')
test = [
    "pytest-cov>=4.1",
    "mypy==1.13.0",
    "pydantic[mypy]>=2.4",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "prp.VERSION"}

[tool.isort]
profile = "black"

[tool.pylint]
max-line-length = 88

[tool.pytest.ini_options]
addopts = "--cov --cov-report html --cov-report term-missing --cov-fail-under 95"

[tool.coverage.run]
source = ["prp"]

[tool.pylint.TYPECHECK]
generated-members = "pysam.asTuple,pysam.TabixFile,pysam.AlignmentFile"

[tool.mypy]
python_version = "3.11"
plugins = "pydantic.mypy"
warn_unused_ignores = true
disallow_untyped_defs = true
no_implicit_optional = true